Implementation of the first step-by-step migration from JSON to Rust storage
This migration synchronizes the JSON and Rust storage at startup. It
stores a JSON checksum in Rust storage to determine if the Rust storage
has diverged (this is the case when this version is started for the
first time, and if the logins for this profile have been processed by
another Firefox in the meantime). Normally, the observer takes over
synchronization so that no migration is necessary on subsequent
launches).

Also: Bug 1970605 - Add JSON storage checksum function to determine rolling migration checkpoint

As part of the AS Login Project, we will have parallel stores for a
while to ensure data integrity in the new store: the JSON store as the
primary store and the new Rust store as a write-only mirror.

To do this, we will implement a rolling (continuous) migration to mirror
the data. However, since a profile can also be opened from a different
Firefox (e.g., a release version if the first FX is a nightly build), it
is conceivable that logins could be changed with an FX that has not been
migrated. To detect this scenario, we will store a checksum of the login
JSON in the Rust store and compare it when the logins are first
accessed.

This patch adds the corresponding functionality to the
JSONFile API, based on IOUtils, which does this in a performant
file-based manner.

Co-authored-by: Johannes Schmidt <joschmidt@mozilla.com>

Differential Revision: https://phabricator.services.mozilla.com/D259801